Export helpful workspace root function
